Separate base and derived classes into files
